move redirect check to protected route
